refactor & rename / tidy up
