+++ src/components/Images/Icons/TwitterIcon.js
@@ -1,17 +1,17 @@
 import React from "react";
 
-const TwitterIcon = ({ fillColor }) => (
-  <svg
-    height="17"
-    viewBox="0 0 20 17"
-    width="20"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="m19.6414599 1.99725146c-.443455.20086842-.9037957.35933626-1.3761558.47406141.5204867-.50522807.9118249-1.13716082 1.1326034-1.8425117.0630171-.2013655-.1510948-.37549123-.3290997-.26767544-.657129.3981579-1.3649635.69685088-2.103893.88787719-.042871.0110848-.0871046.01670176-.1314841.01670176-.1345986 0-.2654502-.05104971-.3683699-.14370468-.78618-.70778655-1.7973236-1.09759357-2.8471046-1.09759357-.4542579 0-.9128954.07232456-1.3632117.21498538-1.3952311.44204971-2.47177613 1.61768421-2.80953769 3.06815205-.12671533.54404971-.1610219 1.08859649-.10204379 1.61847953.00676399.06089181-.0214112.10339181-.03878346.12332456-.03051095.03494445-.07377129.05497661-.11868613.05497661-.00501216 0-.01017032-.00024854-.0152798-.00074561-3.05411193-.28984503-5.80817519-1.7928538-7.75489052-4.23210527-.09927007-.12441812-.28963503-.10915789-.36875912.02947661-.38121655.66807018-.5826764 1.43217544-.5826764 2.20970176 0 1.19159064.4686618 2.31453508 1.28272506 3.13316959-.34233576-.08291228-.67367396-.21349415-.98145985-.38811696-.14890511-.08450293-.33211679.02410819-.33420925.19808479-.02131387 1.77018714.99328467 3.34532164 2.51318735 4.05459943-.03060827.0007456-.06121654.0010936-.09187348.0010936-.24092457 0-.48486618-.0237603-.7249635-.0706345-.16768857-.0327076-.30856448.1297865-.25649635.2958596.49333333 1.5732953 1.81946472 2.7329737 3.40642335 2.9962749-1.31708029.9031374-2.85085158 1.379535-4.44720194 1.379535l-.49766423-.0002982c-.15333334 0-.28301704.1020994-.32209246.253807-.03849149.1494211.03240876.3070439.16316301.3847369 1.79883212 1.0693596 3.85065694 1.634535 5.93454988 1.634535 1.82408759 0 3.53036497-.3697748 5.07148422-1.0989853 1.412798-.6685176 2.6612652-1.6241462 3.7107056-2.8403421.9776155-1.1329357 1.7420924-2.4407924 2.2721168-3.88718425.5052554-1.37873977.7723114-2.84988597.7723114-4.25442398v-.06690643c-.0000487-.22552339.0998053-.43772515.2740146-.58217544.6610705-.54827485 1.2367396-1.19367836 1.7109489-1.91826608.1252068-.19132456-.0722628-.43111404-.2782968-.33776316z"
-      fill={fillColor}
-    />
-  </svg>
-);
-
-export default TwitterIcon;
+export default function TwitterIcon({ fillColor }) {
+  return (
+    <svg
+      height="17"
+      viewBox="0 0 20 17"
+      width="20"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="m19.6414599 1.99725146c-.443455.20086842-.9037957.35933626-1.3761558.47406141.5204867-.50522807.9118249-1.13716082 1.1326034-1.8425117.0630171-.2013655-.1510948-.37549123-.3290997-.26767544-.657129.3981579-1.3649635.69685088-2.103893.88787719-.042871.0110848-.0871046.01670176-.1314841.01670176-.1345986 0-.2654502-.05104971-.3683699-.14370468-.78618-.70778655-1.7973236-1.09759357-2.8471046-1.09759357-.4542579 0-.9128954.07232456-1.3632117.21498538-1.3952311.44204971-2.47177613 1.61768421-2.80953769 3.06815205-.12671533.54404971-.1610219 1.08859649-.10204379 1.61847953.00676399.06089181-.0214112.10339181-.03878346.12332456-.03051095.03494445-.07377129.05497661-.11868613.05497661-.00501216 0-.01017032-.00024854-.0152798-.00074561-3.05411193-.28984503-5.80817519-1.7928538-7.75489052-4.23210527-.09927007-.12441812-.28963503-.10915789-.36875912.02947661-.38121655.66807018-.5826764 1.43217544-.5826764 2.20970176 0 1.19159064.4686618 2.31453508 1.28272506 3.13316959-.34233576-.08291228-.67367396-.21349415-.98145985-.38811696-.14890511-.08450293-.33211679.02410819-.33420925.19808479-.02131387 1.77018714.99328467 3.34532164 2.51318735 4.05459943-.03060827.0007456-.06121654.0010936-.09187348.0010936-.24092457 0-.48486618-.0237603-.7249635-.0706345-.16768857-.0327076-.30856448.1297865-.25649635.2958596.49333333 1.5732953 1.81946472 2.7329737 3.40642335 2.9962749-1.31708029.9031374-2.85085158 1.379535-4.44720194 1.379535l-.49766423-.0002982c-.15333334 0-.28301704.1020994-.32209246.253807-.03849149.1494211.03240876.3070439.16316301.3847369 1.79883212 1.0693596 3.85065694 1.634535 5.93454988 1.634535 1.82408759 0 3.53036497-.3697748 5.07148422-1.0989853 1.412798-.6685176 2.6612652-1.6241462 3.7107056-2.8403421.9776155-1.1329357 1.7420924-2.4407924 2.2721168-3.88718425.5052554-1.37873977.7723114-2.84988597.7723114-4.25442398v-.06690643c-.0000487-.22552339.0998053-.43772515.2740146-.58217544.6610705-.54827485 1.2367396-1.19367836 1.7109489-1.91826608.1252068-.19132456-.0722628-.43111404-.2782968-.33776316z"
+        fill={fillColor}
+      />
+    </svg>
+  );
+}
