+++ src/components/Images/Icons/WhatsappIcon.js
@@ -1,15 +1,17 @@
 import React from "react";
 
-export default ({ fillColor }) => (
-  <svg
-    height="22"
-    viewBox="0 0 22 22"
-    width="22"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="m11 0c-6.06530435 0-11 4.93469565-11 11 0 1.893913.48926087 3.7562609 1.41708696 5.3995652l-1.39939131 4.9930435c-.04591304.1640435-.00143478.3400435.11717392.462.09134782.0942174.21569565.1453913.3433913.1453913.03826087 0 .077-.0047826.11478261-.0138696l5.21113043-1.2908261c1.59117392.854174 3.38321739 1.3046957 5.19582609 1.3046957 6.0653043 0 11-4.9346957 11-11 0-6.06530435-4.9346957-11-11-11zm5.5334783 14.8815652c-.2353044.6513913-1.364 1.2458696-1.9063479 1.3257391-.4868695.0712609-1.1028695.1018696-1.7791304-.1104782-.4098696-.1291304-.9359565-.3003478-1.6098261-.5877826-2.83273912-1.208087-4.68265216-4.0245652-4.82421738-4.2106087-.14108695-.1860435-1.15308695-1.51178263-1.15308695-2.88391306 0-1.37213044.72934782-2.04695652.98856521-2.32626087s.56482609-.34913044.75326087-.34913044.37639131.00239131.54139131.01004348c.17360869.0086087.40652174-.06552174.63560869.47873913.23530435.5586087.80013044 1.93073913.86995652 2.07086957.07078261.13965217.11765218.30273913.02391305.48878261s-.1406087.30226087-.28217392.46534782c-.14156521.16308696-.29652173.36347826-.42373913.48878261-.14156521.13917392-.2883913.28982605-.12386956.56913045.16452174.2793043.73126087 1.1923043 1.57108696 1.9316956 1.07847823.9498261 1.98860873 1.2444348 2.27078263 1.384087s.4471739.1162174.6116956-.0698261c.1645218-.1865217.7059131-.8144783.8938696-1.0933043.1879565-.2788261.3763913-.2329131.6356087-.1396522.2592174.0927826 1.6476087.7671304 1.9297826.9067826s.4706087.2094783.5413913.3256956c.0707826.1157392.0707826.6743479-.1645217 1.3252609z"
-      fill={fillColor}
-    />
-  </svg>
-);
+export default function WhatsappIcon({ fillColor }) {
+  return (
+    <svg
+      height="22"
+      viewBox="0 0 22 22"
+      width="22"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="m11 0c-6.06530435 0-11 4.93469565-11 11 0 1.893913.48926087 3.7562609 1.41708696 5.3995652l-1.39939131 4.9930435c-.04591304.1640435-.00143478.3400435.11717392.462.09134782.0942174.21569565.1453913.3433913.1453913.03826087 0 .077-.0047826.11478261-.0138696l5.21113043-1.2908261c1.59117392.854174 3.38321739 1.3046957 5.19582609 1.3046957 6.0653043 0 11-4.9346957 11-11 0-6.06530435-4.9346957-11-11-11zm5.5334783 14.8815652c-.2353044.6513913-1.364 1.2458696-1.9063479 1.3257391-.4868695.0712609-1.1028695.1018696-1.7791304-.1104782-.4098696-.1291304-.9359565-.3003478-1.6098261-.5877826-2.83273912-1.208087-4.68265216-4.0245652-4.82421738-4.2106087-.14108695-.1860435-1.15308695-1.51178263-1.15308695-2.88391306 0-1.37213044.72934782-2.04695652.98856521-2.32626087s.56482609-.34913044.75326087-.34913044.37639131.00239131.54139131.01004348c.17360869.0086087.40652174-.06552174.63560869.47873913.23530435.5586087.80013044 1.93073913.86995652 2.07086957.07078261.13965217.11765218.30273913.02391305.48878261s-.1406087.30226087-.28217392.46534782c-.14156521.16308696-.29652173.36347826-.42373913.48878261-.14156521.13917392-.2883913.28982605-.12386956.56913045.16452174.2793043.73126087 1.1923043 1.57108696 1.9316956 1.07847823.9498261 1.98860873 1.2444348 2.27078263 1.384087s.4471739.1162174.6116956-.0698261c.1645218-.1865217.7059131-.8144783.8938696-1.0933043.1879565-.2788261.3763913-.2329131.6356087-.1396522.2592174.0927826 1.6476087.7671304 1.9297826.9067826s.4706087.2094783.5413913.3256956c.0707826.1157392.0707826.6743479-.1645217 1.3252609z"
+        fill={fillColor}
+      />
+    </svg>
+  );
+}
